intelligence/requirements.txt
telethon>=1.36
pyyaml>=6.0
pydantic>=2.5
pydantic-settings>=2.2
python-dotenv>=1.0
# LLM (optional — only needed if summarizer is enabled)
emergentintegrations
